anonyme task update

+++ src/Controller/TaskController.php
@@ -3,6 +3,7 @@
 namespace App\Controller;
 
 use App\Entity\Task;
+use App\Entity\User;
 use App\Form\TaskType;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -15,7 +16,23 @@ class TaskController extends AbstractController
     #[Route('/tasks', name: 'task_list')]
     public function list(EntityManagerInterface $entityManager): Response
     {
-        $tasks = $entityManager->getRepository(Task::class)->findBy(['user' => $this->getUser()]);
+        $currentUser = $this->getUser();
+        $taskRepo = $entityManager->getRepository(Task::class);
+
+            $tasks = $taskRepo->findBy(['user' => $currentUser]);
+
+        if ($this->isGranted('ROLE_ADMIN')) {
+            $anonymousUser = $entityManager->getRepository(User::class)->findOneBy(['username' => 'anonyme']);
+            if ($anonymousUser) {
+                $anonymousTasks = $taskRepo->findBy(['user' => $anonymousUser]);
+                $byId = [];
+                foreach (array_merge($tasks, $anonymousTasks) as $t) {
+                    $byId[$t->getId()] = $t;
+                }
+                $tasks = array_values($byId);
+            }
+        }
+
         return $this->render('task/list.html.twig', ['tasks' => $tasks]);
     }
 
@@ -81,6 +98,22 @@ class TaskController extends AbstractController
     #[Route('/tasks/{id}/delete', name: 'task_delete')]
     public function deleteTask(Task $task, EntityManagerInterface $entityManager): Response
     {
+        $currentUser = $this->getUser();
+        if (!$currentUser) {
+            $this->addFlash('error', 'Vous devez être connecté pour supprimer une tâche.');
+            return $this->redirectToRoute('login');
+        }
+
+        $taskOwner = $task->getUser();
+        $isOwner = $taskOwner && $taskOwner->getId() === $currentUser->getId();
+        $isAdmin = $this->isGranted('ROLE_ADMIN');
+        $isAnonymousTask = $taskOwner && $taskOwner->getUsername() === 'anonyme';
+
+        if (!$isOwner && !($isAdmin && $isAnonymousTask)) {
+            $this->addFlash('error', "Vous n'avez pas les droits pour supprimer cette tâche.");
+            return $this->redirectToRoute('task_list');
+        }
+
         $entityManager->remove($task);
         $entityManager->flush();
